Fix the assembly file for Mac
icu_use_data_file_flag is the default so that this file is not used unless icu_use_data_file_flag is set to 0. Just in case, I'm fixing this file.

A fix is a one liner (one character long). '_' was missing in the previous check-in.

@@ -5,7 +5,7 @@
        .data
        .const
        .align 4
-icudt46_dat:
+_icudt46_dat:

R=mark@chromium.org, torne@chromium.org
TBR=mark@chromium.org
BUG=NONE
TEST=With icu_use_data_file_flag=0 in GYPDEFINES, Mac build succeeds.

Review URL: https://codereview.chromium.org/206273002

git-svn-id: http://src.chromium.org/svn/trunk/deps/third_party/icu46@258359 4ff67af0-8c30-449e-8e8b-ad334ec8d88c